requirements.txt
appdirs==1.4.3
astroid==1.4.9
autopep8==1.3.1
backports.functools-lru-cache==1.3
beautifulsoup4==4.5.3
click==6.7
configparser==3.5.0
Flask==0.12
isort==4.2.5
itsdangerous==0.24
JCC==2.22
Jinja2==2.9.5
lazy-object-proxy==1.2.2
lucene==6.2.0
MarkupSafe==0.23
mccabe==0.6.1
mysql-connector-python==2.0.4
nltk==3.2.2
numpy==1.12.0
packaging==16.8
pandas==0.19.2
pycodestyle==2.3.1
pylint==1.6.5
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
url-normalize==1.2.1
virtualenv==15.1.0
Werkzeug==0.11.15
wrapt==1.10.8
